backend/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-adk>=1.26.0",
    "google-cloud-bigquery>=3.40.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "matplotlib>=3.10.8",
    "mcp>=1.26.0",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "toolbox-core>=0.6.0",
    "uvicorn>=0.41.0",
]
